+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add new `--max-time <seconds>` option
+- Add new `--ignore-error` option
+
 ## [0.2.0] - 2022-02-07
 
 ### Added